requirements.txt
pandas
numpy
scikit-learn==1.6.1
lightgbm
shap
fastapi
uvicorn[standard]
streamlit
joblib
python-dotenv
